+++ scripts/build-audit.sh
@@ -119,6 +119,8 @@ FINGERPRINTS='
 0031|wide|lib/wine/x86_64-windows/comdlg32.dll|FileDialogPortal
 0032|ascii|lib/wine/x86_64-windows/libusb-1.0.dll|libusb_submit_transfer
 0033|ascii|lib/wine/x86_64-unix/ntdll.so|WINE_DISABLE_UNIX_MOUNT_REPARSE
+0035|ascii|lib/wine/x86_64-windows/wined3d.dll|Arc(tm) B580
+0036|wide|lib/wine/x86_64-windows/dxgi.dll|__wine_dcomp_null_device
 pipeasio/0001|ascii|lib/wine/x86_64-unix/pipeasio64.dll.so|pipeasio-clamp-sample-rate
 pipeasio/0002|ascii|lib/wine/x86_64-unix/pipeasio64.dll.so|pipeasio-midi-timebase
 '